# rapidapi_service/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Web Metadata & Contact Extractor API",
    "description": "Ultra-fast, enterprise REST API: Extended Metadata, Anti-SSRF Shield, Async Redis Rate Limiter (auto-reconnect + health observability), Multi-Worker Gunicorn, Favicon HD Engine, NLP Summary, HTTP/2, ORJSON.",
    "version": "4.1.0"
  },
  "paths": {
...
    "/api/v1/extract": {
      "get": {
        "tags": [
          "Full Extractor"
        ],
        "summary": "Extract Metadata",
        "description": "Extract full metadata payload (<200ms with Rust ORJSON + IP-Pinned Anti-SSRF Shield + HTTP/2 streaming):\n- **SEO Metadata**: Title, description, OG image, favicon, canonical URL, language, author, theme color, H1 tags.\n- **Page Health Metrics**: Content length (bytes), image count, accessibility missing alt count, link count.\n- **Contacts**: Public email addresses and telephone numbers.\n- **Social Links**: Profiles on Twitter/X, LinkedIn, Instagram, Facebook, GitHub, YouTube, Telegram, TikTok.\n- **Technologies**: Precision context-aware CMS and framework signatures.\n- **SEO Audit**: Automated 14-point SEO diagnostic score & warnings.\n- **Link Extractor**: Categorized internal vs external hyperlinks.\n- **Structured Data**: Schema.org JSON-LD schemas.\n- **Feeds**: RSS/Atom feed discovery.\n- **Security**: HTTP security headers score & IP-Pinned Anti-SSRF Protection.\n- **AI Reader**: Clean Markdown article text, word count & reading time.\n\nWithout `fields`, runs the complete pipeline. With `fields`, only runs\nthe extractors needed for the requested keys — `fields=detected_technologies`\nis exactly as cheap as `GET /api/v1/tech-stack`.",
        "operationId": "extract_metadata_api_v1_extract_get",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The target website URL to analyze (e.g. https://example.com)",
              "title": "Url"
            },
            "description": "The target website URL to analyze (e.g. https://example.com)"
          },
          {
            "name": "fields",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional comma-separated list of keys to filter response (e.g. metadata,contacts) — only runs the extractors those fields actually need, same as calling the matching specialized endpoint",
              "title": "Fields"
            },
            "description": "Optional comma-separated list of keys to filter response (e.g. metadata,contacts) — only runs the extractors those fields actually need, same as calling the matching specialized endpoint"
          },
          {
            "name": "user_agent",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional custom User-Agent header",
              "title": "User Agent"
            },
            "description": "Optional custom User-Agent header"
          },
          {
            "name": "x-rapidapi-proxy-secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Rapidapi-Proxy-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad Request: Invalid URL format, missing domain hostname, or blocked by IP-Pinned Anti-SSRF Shield"
          },
          "403": {
            "description": "Forbidden: Invalid or missing X-RapidAPI-Proxy-Secret authentication header"
          },
          "429": {
            "description": "Too Many Requests: Rate limit exceeded (60 requests/min per IP)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tech-stack": {
      "get": {
        "tags": [
          "Tech Stack"
        ],
        "summary": "Extract Tech Stack",
        "description": "Dedicated endpoint for technology intelligence & CMS auditing:\nDetects 40+ frameworks and CMS signatures with structural context-aware matching.",
        "operationId": "extract_tech_stack_api_v1_tech_stack_get",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The target URL to inspect for CMS and technology stack signatures",
              "title": "Url"
            },
            "description": "The target URL to inspect for CMS and technology stack signatures"
          },
          {
            "name": "user_agent",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional custom User-Agent header",
              "title": "User Agent"
            },
            "description": "Optional custom User-Agent header"
          },
          {
            "name": "x-rapidapi-proxy-secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Rapidapi-Proxy-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechStackResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request: Invalid URL format, missing domain hostname, or blocked by IP-Pinned Anti-SSRF Shield"
          },
          "403": {
            "description": "Forbidden: Invalid or missing X-RapidAPI-Proxy-Secret authentication header"
          },
          "429": {
            "description": "Too Many Requests: Rate limit exceeded (60 requests/min per IP)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TechStackResponse": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "final_url": {
            "type": "string",
            "title": "Final Url"
          },
          "status_code": {
            "type": "integer",
            "title": "Status Code"
          },
          "execution_time_ms": {
            "type": "number",
            "title": "Execution Time Ms"
          },
          "bot_protection_detected": {
            "type": "boolean",
            "title": "Bot Protection Detected",
            "default": false
          },
          "detected_technologies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Detected Technologies"
          },
          "technology_details": {
            "items": {
              "$ref": "#/components/schemas/TechnologyDetail"
            },
            "type": "array",
            "title": "Technology Details",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "url",
          "final_url",
          "status_code",
          "execution_time_ms",
          "detected_technologies"
        ],
        "title": "TechStackResponse"
      },
      "TechnologyDetail": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence",
            "default": []
          },
          "category": {
            "type": "string",
            "title": "Category"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "name",
          "confidence",
          "category"
        ],
        "title": "TechnologyDetail"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}